Piccoli passi avanti

+++ analog.h
@@ -7,16 +7,16 @@
 #define _ANALOG_H
 #include "common.h"
 
-#define SP_TH1_COUNT 256 // TBD
-#define SP_TH2_COUNT 512 // TBD
-#define SP_TH3_COUNT 1023 // TBD
+#define SP_TH1_COUNT 256 
+#define SP_TH2_COUNT 512
+#define SP_TH3_COUNT 1023
 
 
 typedef struct
 {
 	
-	unsigned short correnteCarica;
-	unsigned short distanceSetPoint;
+	unsigned int correnteCarica;
+	unsigned int distanceSetPoint;
 } tAnalog;
 
 void analogInit(void);
